pyproject.toml
[project]
name = "clawmon"
version = "0.1.0"
description = "OpenClaw exposure monitoring scanner"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
